+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+Version 0.2
+===========
+
+- switch back to default ``repr`` implementation
+- have ``template`` as singleton attribute
+- some more docs
+
 Version 0.1
 ===========